starting to split api routes into separate files
